/scheduler endpoint -> /status, /schedule gets its own app
Also, learned about django-pytest's "transactional" db option and some
shared-cache (cross-thread db connection)  support that only shows up in py3's
version of pysqlite, so a test case I had written for the new .status property
on the scheduler was very finicky, and you cans see for of that finickiness in
the modification to test_scheduler.py. I think that these issues are mostly
related to the in-memory sqlite db used during testing, so I will not spend TOO
much more time on trying to get multi-threaded test cases humming.