Add ability to see USRP USB device in docker container. (#1)
* Add ability to see USRP USB device in docker container.

* Update README.md

* Update README.md

* Docker apt install bug (#2)

* Fix network error in apt installs inside docker.

Adding multiple calls that install `gnuradio`'s and `uhd-host`'s dependencies sidesteps the bad hashes we were previously experiencing due to our network configuration. This is ugly, and should probably be refined later, but at least it works now.

* Add `--build` flag to docker-compose.yml.

This forces a rebuild upon source changes.

* Add RUN command for UHD image script.

* Add ability to see USRP USB device in docker container.

* Add different path to `devices` key in docker-compose file.
In order to mount X USB devices to the `api` container, we had to list `/dev/bus/usb` as the source and destination in the compose file.
This has to be specified to run in `rw` mode in order to avoid the `priviledged` flag.