Get the test_scheduler tests passing. (#5)
* Get the test_scheduler tests passing.
The problem is in the `create_entry` function. ScheduleEntry objects now need assigned users, which at not allowed to be NULL. To compensate for this dependency, create a global user who will own all of the tested entries. Then put the user under the `kwargs` dic, with the key `'owner'`, in the `create_entry` function. This should solve all failures on this test.

* Create `temp_user` inside of `create_entry`.
Global cannot be used due to marking issues. Instead create a temp var inside of the function.

* Add Doug's suggestion.