Clean up the interaction between gunicorn processes and scheduler thread.
Gunicorn's master process doesn't propagate its signals (SIGTERM, etc) to the
worker process. Instead, it has an internal boolean flag that the worker
watches. We needed to find a way to set the scheduler's own interrupt flag when
it was time to shut down, and for the Gunicorn provides server
hooks (http://docs.gunicorn.org/en/stable/settings.html#server-hooks). By using
the `worker_exit` server hook, we are able to run predefined code within the
worker process's context when the worker is supposed to shut down. Within that
function, we bootstrap django and get a handle to the scheduler thread and
set its interrupt flag.

Closes SMSS-157.