Refactor and test max_results capability.
max_results essentially says that since TaskResults are not manually
administered by users (like Acquisitions are, for example, where the user who
requested an Acquisition is supposed to retreive it and delete it in a
reasonable amount of time), TaskResult's save method is overridden to keep the
total number of results below a threshold set in the global settings.py file.
The algorithm used is, if during new task result creation, the object would be
the (MAX_TEST_RESULTS + 1)th result, then the oldest result is deleted, and
then the new result is saved.