Implement `validate_only` option on ScheduleEntry serializer
The goal here it to allow users POST a schedule entry and get back info on
whether it's valid or not immediately without modifying the schedule.

Whereas normal posting (`validate_only=False`) returns 201 CREATED on success,
when `validate_only=True` we must return 204 NO CONTENT (rather than 200 OK)
since we cannot provide the normal hyperlinking that an existing schedule entry
has.

It still responds with 400 BAD REQUEST if invalid, whether `validate_only` is
True or not.

The use case for this option is if a system like a sensor manager takes
schedule "requests" from clients and then an "owner" has to allow the request,
may a day or more later, there needs to be a more instant feedback mechanism
for at the time the request is being made without modifying the schedule.