Floor division of a negative number wrong result
Should not use `int(longitude // 100)` but rather `int(longitude / 100)`. When the longitude is west of the prime meridian, the longitude is written as a negative number. Floor division (`//`)  of a negative number does not give the same result as flooring a positive number and then adding a negative sign:

```
>>> (-1234.5678 // 100)
-13.0
>>> -(1234.5678 // 100)
-12.0
```

Also changing latitude calculation for consistency / potential southern hemisphere issue.