Changes Task "time" serialization from timestamp to datetime string
Also, cap the number of tasks diplayed in /status to settings.MAX_TASK_QUEUE,
since there is a potential edge case where, if all tasks are created with
identical intervals, as the number of entries in the schedule grows, the number
of tasks in the task queue grows without bound. The /status task queue is
intended to be a snapshot of upcoming tasks, not an exhaustive list of all
tasks.